// src/Models/Contract/WarehouseModel.ts
export type WarehouseModel = {
  warehouse_name: string;
  in_warehouse: number;
  sku_ended: number;
  balance_cost: number;
  revenue: number;
  in_way_to_client: number;
  in_way_from_client: number;
  redemption_percentage: number;
  products_delivered: number;
  last_delivery: string;
  storage_cost: number;
  income_id: number;
  sku: number;
};